pagination done

+++ util/file.js
@@ -1,6 +1,11 @@
 const fs = require('fs');
 
 const deleteFile = (filePath) => {
+  if(filePath.match(/^(http|https):\/\//)) {
+    // It's an external URL, do not attempt to delete
+    console.log('External file URL, skipping deletion:', filePath);
+    return;
+  }
   fs.unlink(filePath, (err) => {
     if (err) { throw err; } 
     else { console.log('File deleted successfully'); }
